address Pani's comments
