+++ integration-tests/pom.xml
@@ -356,7 +356,7 @@
                 <includes-failsafe>
                     **/ItKubernetesEvents,
                     **/ItMiiAuxiliaryImage,
-                    **/ItParameterizedDomainNoLB,
+                    **/ItParameterizedDomain,
                     **/ItPodsRestart,
                     **/ItTwoDomainsManagedByTwoOperators
                 </includes-failsafe>
